tools: Bugfix converting VariantContext to Variant. Missing FORMAT fields.
Because HTSJDK is not saving the original FORMAT the only way to do that
is reading, from one sample, all the fields that has. It may happen
that the first, or many samples have missing values.

The current algorithm was reading the existing fields in the first sample.
It was not storing correctly the FORMAT list if the first sample has
missing values.

 Changes not staged for commit: